added html/css/json
In this pull req/commit, I created our home page and put the elements that were originaly the index page on the activities page, because they had all of the requirements for the activities page. I added images, a form, and more text to the html. I added some new styling to the css, and I put text in the json files and added some effects